allow redirecting to the first page when an invalid page is requested
When the number of items in a collection varies, it happens that urls,
already shared on the web, point to a page that is now invalid.
Up to now dj-pagination only allowed to either render nothing or raise
a 404 page. Another  good practice to prevent broken links is to
redirect to the first page with a 301 http code.